requirements.txt
click==8.1.7
pytest==7.4.3
PyYAML==6.0.1
SQLAlchemy==2.0.22
